requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Django==4.2.3
executing==1.2.0
filelock==3.12.2
fonttools==4.40.0
idna==3.4
image==1.5.33
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
supervision==0.11.1
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.132
urllib3==2.0.3
wcwidth==0.2.6
